Updated entities and added global exception concept

+++ E_CommerceApp/src/main/java/com/My/E_CommerceApp/DTO/RequestDTO/PaymentRequestDTO.java
@@ -2,9 +2,11 @@ package com.My.E_CommerceApp.DTO.RequestDTO;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PaymentRequestDTO {
-    private Long orderId;          // যে Order এর জন্য payment হবে তার ID
-    private Long amount;     // মোট টাকা (Order এর মোট amount)
+    private Long orderId;
+    private BigDecimal amount;
     private String paymentMethod;
 }
